LayoutTests/imported/w3c/web-platform-tests/screen-wake-lock/wakelock-enabled-by-feature-policy-attribute-redirect-on-load.https.sub-expected.txt

CONSOLE MESSAGE: Feature policy 'ScreenWakeLock' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute ';screen-wake-lock'.

PASS Feature-Policy allow="screen-wake-lock" allows same-origin relocation
PASS Feature-Policy allow="screen-wake-lock" disallows cross-origin relocation
